+++ Changelog.md
@@ -19,6 +19,8 @@ Bugfixes:
  * SMTChecker: Fix internal error on pushing to ``string`` casted to ``bytes``.
 
 AST Changes:
+ * ModifierInvocation: Add ``kind`` field which can be ``modifierInvocation`` or ``baseConstructorSpecifier``.
+
 
 ### 0.8.2 (2021-03-02)